Refactoring: DefaultSubmissionClient / ISubmissionAdapter
