+++ src/configuration/IConfigurationSettings.ts
@@ -3,10 +3,10 @@ import { ILog } from '../logging/ILog';
 import { IEventQueue } from '../queue/IEventQueue';
 import { IEnvironmentInfoCollector } from '../services/IEnvironmentInfoCollector';
 import { IErrorParser } from '../services/IErrorParser';
-import { IExitController } from '../services/IExitController';
 import { IModuleCollector } from '../services/IModuleCollector';
 import { IRequestInfoCollector } from '../services/IRequestInfoCollector';
 import { IStorage } from '../storage/IStorage';
+import { ISubmissionAdapter } from '../submission/ISubmissionAdapter';
 import { ISubmissionClient } from '../submission/ISubmissionClient';
 
 export interface IConfigurationSettings {
@@ -20,7 +20,7 @@ export interface IConfigurationSettings {
   requestInfoCollector?:IRequestInfoCollector;
   submissionBatchSize?:number;
   submissionClient?:ISubmissionClient;
+  submissionAdapter?:ISubmissionAdapter;
   storage?:IStorage<any>;
   queue?:IEventQueue;
-  exitController?: IExitController;
 }
