--- src/services/IExitController.ts
@@ -1,6 +1,0 @@
-import { Configuration } from '../configuration/Configuration';
-
-export interface IExitController {
-  isApplicationExiting:boolean;
-  processExit(config:Configuration):void;
-}
