+++ src/submission/ISubmissionAdapter.ts
@@ -1,0 +1,6 @@
+import { SubmissionCallback } from './SubmissionCallback';
+import { SubmissionRequest } from './SubmissionRequest';
+
+export interface ISubmissionAdapter {
+  sendRequest(request:SubmissionRequest, callback:SubmissionCallback): void
+}
