+++ src/submission/SubmissionCallback.ts
@@ -1,0 +1,3 @@
+export interface SubmissionCallback {
+  (status: number, message: string, data?: string, headers?: Object): void
+}
