+++ src/submission/SubmissionRequest.ts
@@ -1,0 +1,8 @@
+export interface SubmissionRequest {
+  serverUrl: string;
+  apiKey: string;
+  userAgent: string;
+  method: string;
+  path: string;
+  data: string;
+}
